Moved mining features into one package
